requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.1
pgvector==0.3.6

# Auth
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.19

# HTTP client
httpx==0.28.1

# LLM / AI
google-genai==1.33.0

# ML
scikit-learn==1.6.1
joblib==1.4.2
numpy==2.2.1

# Validation
jsonschema==4.23.0

# Utilities
python-dateutil==2.9.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.1
